Employee System Integration Format/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
python-dotenv
openai
pydantic
SQLAlchemy
psycopg2-binary
redis
fuzzywuzzy
thefuzz
python-Levenshtein
bcrypt
requests
pandas
openpyxl